Deduplicate function zero-grad.
It was an issue when running code in notebooks
